move notzero test
